receipt-analyzer/requirements.txt
pillow==10.0.0
opencv-python-headless==4.8.0.76
pytesseract==0.3.11
pandas==2.2.0
openpyxl==3.1.2
pyyaml==6.0
pydantic==2.3.0
python-dotenv==1.0.0
pytest==7.4.0
